Refactor animated text from spinner view
